UPDATES: RI/Sites changes and Segments fix
Segments: It was not possible to implement a more efficient method in Segments to obtain the parents, because the list returned by the Pythoncyc function does not always come in order. For now it stayed the same but the method will be moved to another site to improve the performance of the code. The left_end_position and right_end_position properties have been renamed to match JSONschema.
RI/Sites: The 'sRNAs' were added to the download of RIs, as well as the new property 'mechanism' in RI as well as in Sites.
No changes needed to be made to leftEndposition and rightEndposition in Sites, they were just added to the set_site method.

Phill Betancourt ☕