TEST: Test module for new releases and debugging.
Test module was added: The module checks functions works for debogging porposes. Also the module can test if there where minor changes in the extracted collections related to the collections in our database based on the IDs registered.
Readme: Badges versions update and orthographics corrections.
Snakefile: Commented code corrected.
Some files has modifications in the import statements for testing porposes, it does not affect the code execution and will be implemented eventually in the rest of the files.

Author: Phill Betancourt ☕